Address review feedback: name-uniqueness resolution fixes
- Exempt only registered default services from machine-wide name uniqueness
  (identity-based, via IsNameUniquenessExempt) rather than any resource
  literally named "builtin"; two non-default resources sharing a name still
  collide, a lone one is allowed.
- Split simple-name lookup: FindBySimpleName resolves to the single owner
  (local wins over same-named remotes) and returns a typed error
  (NodeNotFoundError / MultipleMatchingNamesError); FindAllBySimpleName
  enumerates every match for collision logging. Both the required- and
  optional-dependency resolvers use the winner, fixing the local-vs-remote
  resolution bug in each.
- Simplify collidingLocalNames to group-by-name; comment/nit fixes.
- Move end-to-end collision tests into resource_name_uniqueness_test.go.
