+++ resource/resource_graph.go
@@ -16,8 +16,8 @@ import (
 	"go.viam.com/rdk/logging"
 )
 
-// NodeNotFoundError is returned when the name and API combination are not found after
-// by FindBySimpleNameAndAPI.
+// NodeNotFoundError is returned when the name and API combination are not found by
+// FindBySimpleNameAndAPI, or when a simple name is not found by FindBySimpleName.
 type NodeNotFoundError struct {
 	Name string
 	API  API
@@ -50,6 +50,24 @@ func (e *MultipleMatchingRemoteNodesError) Error() string {
 func IsMultipleMatchingRemoteNodesError(err error) bool {
 	var mmrnErr *MultipleMatchingRemoteNodesError
 	return errors.As(err, &mmrnErr)
+}
+
+// MultipleMatchingNamesError is returned by FindBySimpleName when a simple name is claimed by more
+// than one resource (multiple locals, or multiple remotes with no local) so no single owner can be
+// resolved under machine-wide name uniqueness. Matches holds every colliding name.
+type MultipleMatchingNamesError struct {
+	Name    string
+	Matches []Name
+}
+
+func (e *MultipleMatchingNamesError) Error() string {
+	return fmt.Sprintf("found multiple resources matching name %q: %v", e.Name, NamesToStrings(e.Matches))
+}
+
+// IsMultipleMatchingNamesError returns if the given error is any kind of multiple matching names error.
+func IsMultipleMatchingNamesError(err error) bool {
+	var mmnErr *MultipleMatchingNamesError
+	return errors.As(err, &mmnErr)
 }
 
 type graphNodes map[Name]*GraphNode
@@ -428,7 +446,7 @@ func (g *Graph) SimpleNamesWhere(filter func(Name, *GraphNode) bool) []Name {
 		}
 	}
 
-	// Group candidates by bare name to enforce name uniqueness across APIs.
+	// Group candidates by simple name to enforce name uniqueness across APIs.
 	byName := make(map[string][]candidate)
 	for k, v := range g.nodes.simpleNameCache {
 		if v.local == nil && len(v.remote) != 1 {
@@ -443,11 +461,10 @@ func (g *Graph) SimpleNamesWhere(filter func(Name, *GraphNode) bool) []Name {
 			c.node = remNode
 		}
 
-		// Only components and services participate in machine-wide name uniqueness. The
-		// reserved default-service name and rdk-internal resources are exempt and emitted per-API.
-		if k.name == DefaultServiceName ||
-			k.api.Type.Namespace == APINamespaceRDKInternal ||
-			!(k.api.IsComponent() || k.api.IsService()) {
+		// Only components and services participate in machine-wide name uniqueness. Registered
+		// default services (which share the reserved "builtin" name across service APIs) and
+		// rdk-internal resources are exempt and emitted per-API.
+		if IsNameUniquenessExempt(k.name, k.api) {
 			emit(c)
 			continue
 		}
@@ -515,15 +532,20 @@ func (g *Graph) FindNodesByAPI(api API) []Name {
 	return ret
 }
 
-// FindBySimpleName returns all unmodified resource names that match the
-// provided string after applying any remote prefixes. This means that while
-// the "name" argument to this function should include remote prefix(es), the
-// Name.Name field of the return value: will not include remote prefix(es).
+// FindBySimpleName resolves the given simple name to the single resource that owns it under
+// machine-wide name uniqueness: a local resource wins over same-named remotes. It returns a
+// NodeNotFoundError when nothing matches and a MultipleMatchingNamesError when the name has
+// collisions (more than one local claimant, or no local claimant with more than one remote
+// claimant). The "name" argument should include any remote prefix(es); the Name.Name field
+// of the return value will not.
 //
-// It will also ignore internal resources (if namespace is "rdk-internal") because
-// none of them besides the framesystem have exposed grpc methods. In any case, the framesystem
-// is expected to be called for differently.
-func (g *Graph) FindBySimpleName(name string) []Name {
+// Internal resources ("rdk-internal" namespace) are ignored because none of them besides the
+// framesystem have exposed grpc methods. In any case, the framesystem is expected to be called for
+// differently.
+//
+// Use [Graph.FindAllBySimpleName] instead when every matching name is needed (e.g. to enumerate
+// collisions for logging) rather than the single resolved winner.
+func (g *Graph) FindBySimpleName(name string) (Name, error) {
 	g.mu.RLock()
 	defer g.mu.RUnlock()
 	return g.findBySimpleName(name)
@@ -531,7 +553,49 @@ func (g *Graph) FindBySimpleName(name string) []Name {
 
 // findBySimpleName is the unlocked implementation of [Graph.FindBySimpleName]. Callers must hold
 // g.mu.
-func (g *Graph) findBySimpleName(name string) []Name {
+func (g *Graph) findBySimpleName(name string) (Name, error) {
+	matches := g.namesMatchingSimpleName(name)
+	var locals, remotes []Name
+	for _, n := range matches {
+		if n.Remote == "" {
+			locals = append(locals, n)
+		} else {
+			remotes = append(remotes, n)
+		}
+	}
+	switch {
+	case len(locals) == 1:
+		// A local resource wins its simple name over any same-named remotes.
+		return locals[0], nil
+	case len(locals) == 0 && len(remotes) == 1:
+		return remotes[0], nil
+	case len(matches) == 0:
+		return Name{}, &NodeNotFoundError{Name: name}
+	default:
+		// More than one local claimant, or no local claimant with more than one remote claimant:
+		// the name has no single owner.
+		return Name{}, &MultipleMatchingNamesError{Name: name, Matches: matches}
+	}
+}
+
+// FindAllBySimpleName returns every unmodified resource name that matches the provided string
+// after applying any remote prefixes. The "name" argument should include any remote prefix(es);
+// the Name.Name field of the return value will not. Unlike [Graph.FindBySimpleName] it does not
+// apply machine-wide name-uniqueness resolution, so it may return several names sharing a simple
+// name; use it where all matches are needed (e.g. to enumerate collisions for logging).
+//
+// Internal resources ("rdk-internal" namespace) are ignored because none of them besides the
+// framesystem have exposed grpc methods. In any case, the framesystem is expected to be called for
+// differently.
+func (g *Graph) FindAllBySimpleName(name string) []Name {
+	g.mu.RLock()
+	defer g.mu.RUnlock()
+	return g.namesMatchingSimpleName(name)
+}
+
+// namesMatchingSimpleName is the unlocked all-matches scan backing [Graph.FindAllBySimpleName].
+// Callers must hold g.mu.
+func (g *Graph) namesMatchingSimpleName(name string) []Name {
 	var result []Name
 	for key, val := range g.nodes.simpleNameCache {
 		if key.api.Type.Namespace == APINamespaceRDKInternal {
@@ -904,32 +968,32 @@ func (g *Graph) ResolveDependencies(logger logging.Logger) error {
 
 				// if a name is later added that conflicts, it will not
 				// necessarily be caught unless the resource config changes.
-				nodeNames := g.findBySimpleName(dep)
-				switch len(nodeNames) {
-				case 0:
-				case 1:
-					if nodeNames[0].String() == nodeName.String() {
-						allErrs = multierr.Combine(errors.Errorf("node cannot depend on itself: %q", nodeName))
-						logger.Errorw("node cannot depend on itself", "name", nodeName)
-						return Name{}, false
+				resolved, err := g.findBySimpleName(dep)
+				if err != nil {
+					var multiErr *MultipleMatchingNamesError
+					if errors.As(err, &multiErr) {
+						allErrs = multierr.Combine(
+							allErrs,
+							errors.Errorf("conflicting names for resource %q: %v", nodeName, NamesToStrings(multiErr.Matches)))
+						logger.Errorw(
+							"cannot resolve dependency for resource due to multiple matching names",
+							"name", nodeName,
+							"conflicts", NamesToStrings(multiErr.Matches),
+						)
 					}
-					logger.Debugw(
-						"dependency resolved for resource",
-						"name", nodeName,
-						"dependency", nodeNames[0],
-					)
-					return nodeNames[0], true
-				default:
-					allErrs = multierr.Combine(
-						allErrs,
-						errors.Errorf("conflicting names for resource %q: %v", nodeName, NamesToStrings(nodeNames)))
-					logger.Errorw(
-						"cannot resolve dependency for resource due to multiple matching names",
-						"name", nodeName,
-						"conflicts", NamesToStrings(nodeNames),
-					)
+					return Name{}, false
 				}
-				return Name{}, false
+				if resolved.String() == nodeName.String() {
+					allErrs = multierr.Combine(errors.Errorf("node cannot depend on itself: %q", nodeName))
+					logger.Errorw("node cannot depend on itself", "name", nodeName)
+					return Name{}, false
+				}
+				logger.Debugw(
+					"dependency resolved for resource",
+					"name", nodeName,
+					"dependency", resolved,
+				)
+				return resolved, true
 			}
 
 			resolvedName, resolved := tryResolve()
