+++ robot/impl/optional_dependencies_test.go
@@ -13,14 +13,17 @@ import (
 	"go.viam.com/utils"
 	gotestutils "go.viam.com/utils/testutils"
 
+	"go.viam.com/rdk/components/arm"
 	"go.viam.com/rdk/components/generic"
 	"go.viam.com/rdk/components/motor"
 	"go.viam.com/rdk/components/motor/fake"
+	"go.viam.com/rdk/components/sensor"
 	"go.viam.com/rdk/config"
 	"go.viam.com/rdk/logging"
 	"go.viam.com/rdk/resource"
 	"go.viam.com/rdk/robot"
 	"go.viam.com/rdk/testutils"
+	"go.viam.com/rdk/testutils/inject"
 	"go.viam.com/rdk/testutils/robottestutils"
 	rutils "go.viam.com/rdk/utils"
 )
@@ -960,6 +963,55 @@ func TestModularOptionalDependencyOnRemote(t *testing.T) {
 	doCommandResp, err = fooRes.DoCommand(ctx, map[string]any{"command": "optional_motor_state"})
 	test.That(t, err, test.ShouldBeNil)
 	test.That(t, doCommandResp, test.ShouldResemble, map[string]any{"optional_motor_state": "moving: false"})
+}
+
+func TestOptionalDependencyLocalWinsOverRemote(t *testing.T) {
+	logger := logging.NewTestLogger(t)
+	ctx := context.Background()
+
+	lr := setupLocalRobot(t, ctx, &config.Config{}, logger, WithDisableCompleteConfigWorker()).(*localRobot)
+
+	// A local arm and an (unprefixed) remote sensor both named "shared".
+	injectArm := &inject.Arm{}
+	localArmCfg := resource.Config{API: arm.API, Name: "shared"}
+	err := lr.manager.resources.AddNode(
+		localArmCfg.ResourceName(),
+		resource.NewConfiguredGraphNode(localArmCfg, injectArm, localArmCfg.Model),
+	)
+	test.That(t, err, test.ShouldBeNil)
+
+	injectSensor := &inject.Sensor{}
+	remoteSensorCfg := resource.Config{API: sensor.API, Name: "shared"}
+	err = lr.manager.resources.AddNode(
+		remoteSensorCfg.ResourceName().PrependRemote("remote1"),
+		resource.NewConfiguredGraphNode(remoteSensorCfg, injectSensor, remoteSensorCfg.Model),
+	)
+	test.That(t, err, test.ShouldBeNil)
+
+	// FindAllBySimpleName surfaces both matches (the ambiguity that used to make the resolver bail
+	// out), while FindBySimpleName applies name uniqueness and resolves to the local arm.
+	test.That(t, len(lr.manager.resources.FindAllBySimpleName("shared")), test.ShouldEqual, 2)
+	resolvedShared, err := lr.manager.resources.FindBySimpleName("shared")
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, resolvedShared, test.ShouldResemble, arm.Named("shared"))
+
+	// A resource optionally depending on the bare name "shared".
+	dependentCfg := resource.Config{
+		API:                       generic.API,
+		Name:                      "dependent",
+		ImplicitOptionalDependsOn: []string{"shared"},
+	}
+	optDeps, snapshot := lr.getOptionalDependenciesAndSnapshot(dependentCfg)
+
+	// The optional dependency resolves to the local arm, not the remote sensor.
+	test.That(t, len(optDeps), test.ShouldEqual, 1)
+	resolved, ok := optDeps[arm.Named("shared")]
+	test.That(t, ok, test.ShouldBeTrue)
+	test.That(t, resolved, test.ShouldEqual, injectArm)
+	_, sensorResolved := optDeps[sensor.Named("shared")]
+	test.That(t, sensorResolved, test.ShouldBeFalse)
+	_, inSnapshot := snapshot[arm.Named("shared")]
+	test.That(t, inSnapshot, test.ShouldBeTrue)
 }
 
 func TestModularOptionalDependencyOnRemoteWithPrefix(t *testing.T) {
