+++ robot/impl/resource_name_uniqueness_test.go
@@ -1,0 +1,620 @@
+package robotimpl
+
+import (
+	"context"
+	"errors"
+	"sync/atomic"
+	"testing"
+	"time"
+
+	"go.viam.com/test"
+	goutils "go.viam.com/utils"
+	"go.viam.com/utils/testutils"
+
+	"go.viam.com/rdk/components/arm"
+	"go.viam.com/rdk/components/movementsensor"
+	"go.viam.com/rdk/components/sensor"
+	"go.viam.com/rdk/config"
+	"go.viam.com/rdk/logging"
+	"go.viam.com/rdk/resource"
+	rdktestutils "go.viam.com/rdk/testutils"
+	"go.viam.com/rdk/testutils/inject"
+	"go.viam.com/rdk/testutils/robottestutils"
+	"go.viam.com/rdk/utils"
+)
+
+func TestFullResourceNameCollision(t *testing.T) {
+	// Asserts the following as features of RSDK-11267/RSDK-11268:
+	// - Duplicately-named local resources are not reachable through an SDK
+	// - Duplicately-named remote resources are not reachable through an SDK
+	// - Local resources with the same full name as a remote resource are "preferred"
+	// - Logs for full resource name collisions are correctly generated by the resource
+	//   manager
+	// - Reconfiguring a remote to have a prefix can "solve" a full resource name collision
+
+	// Set up three injected arms and track their IsMoving counts separately.
+	var arm1IsMovingCount, arm2IsMovingCount, arm3IsMovingCount atomic.Int32
+	arm1 := &inject.Arm{
+		IsMovingFunc: func(context.Context) (bool, error) {
+			arm1IsMovingCount.Add(1)
+			return false, nil
+		},
+	}
+	arm2 := &inject.Arm{
+		IsMovingFunc: func(context.Context) (bool, error) {
+			arm2IsMovingCount.Add(1)
+			return false, nil
+		},
+	}
+	arm3 := &inject.Arm{
+		IsMovingFunc: func(context.Context) (bool, error) {
+			arm3IsMovingCount.Add(1)
+			return false, nil
+		},
+	}
+	resetAllIsMovingCounts := func() {
+		arm1IsMovingCount.Store(0)
+		arm2IsMovingCount.Store(0)
+		arm3IsMovingCount.Store(0)
+	}
+
+	// Use a consistent model and constructor for all three arms, but vary the actual arm
+	// created based on a config value. This will help us make assertions on _where_
+	// IsMoving requests are routing to. This is likely redundant with some of the tests
+	// above, but we want a full integration test here.
+	model := resource.DefaultModelFamily.WithModel(goutils.RandomAlphaString(8))
+	resource.RegisterComponent(
+		arm.API,
+		model,
+		resource.Registration[arm.Arm, resource.NoNativeConfig]{Constructor: func(
+			ctx context.Context,
+			deps resource.Dependencies,
+			conf resource.Config,
+			logger logging.Logger,
+		) (arm.Arm, error) {
+			switch conf.Attributes.String("armToRouteTo") {
+			case "arm1":
+				return arm1, nil
+			case "arm2":
+				return arm2, nil
+			case "arm3":
+				return arm3, nil
+			default:
+				return nil, errors.New("unknown armToRouteTo provided")
+			}
+		}},
+	)
+	defer resource.Deregister(arm.API, model)
+
+	logger, logs := logging.NewObservedTestLogger(t)
+	blankLogger := logging.NewBlankLogger("") // To be used where we don't care about logging.
+	ctx := context.Background()
+
+	// Setup three machines.
+	r := setupLocalRobot(t, ctx, &config.Config{}, logger)
+	options, _, raddr := robottestutils.CreateBaseOptionsAndListener(t)
+	err := r.StartWeb(ctx, options)
+	test.That(t, err, test.ShouldBeNil)
+	r2 := setupLocalRobot(t, ctx, &config.Config{}, blankLogger)
+	options, _, r2addr := robottestutils.CreateBaseOptionsAndListener(t)
+	err = r2.StartWeb(ctx, options)
+	test.That(t, err, test.ShouldBeNil)
+	r3 := setupLocalRobot(t, ctx, &config.Config{}, blankLogger)
+	options, _, r3addr := robottestutils.CreateBaseOptionsAndListener(t)
+	err = r3.StartWeb(ctx, options)
+	test.That(t, err, test.ShouldBeNil)
+
+	// Create clients to two of the machines (to be used throughout tests).
+	rClient := robottestutils.NewRobotClient(t, blankLogger, raddr, time.Second)
+	r2Client := robottestutils.NewRobotClient(t, blankLogger, r2addr, time.Second)
+
+	{ // Single local "fooArm" instance.
+		cfg := &config.Config{
+			Components: []resource.Config{
+				{
+					Name:       "fooArm",
+					Model:      model,
+					API:        arm.API,
+					Attributes: utils.AttributeMap{"armToRouteTo": "arm1"},
+				},
+			},
+		}
+		r.Reconfigure(ctx, cfg)
+
+		// Verify no logs were generated complaining about collisions.
+		test.That(t, logs.FilterMessageSnippet("collision").Len(), test.ShouldEqual, 0)
+
+		// Verify (refreshed) resource names returned to rClient contain only one "fooArm" instance.
+		expectedNames := []resource.Name{arm.Named("fooArm")}
+		test.That(t, rClient.Refresh(ctx), test.ShouldBeNil)
+		actualNames := rClient.ResourceNames()
+		rdktestutils.VerifySameResourceNames(t, actualNames, expectedNames)
+
+		// Call IsMoving() through "fooArm" from rClient, and assert that request routed to
+		// arm1.
+		fooArmResClient, err := rClient.ResourceByName(arm.Named("fooArm"))
+		test.That(t, err, test.ShouldBeNil)
+		fooArmClient, ok := fooArmResClient.(arm.Arm)
+		test.That(t, ok, test.ShouldBeTrue)
+		isMoving, err := fooArmClient.IsMoving(ctx)
+		test.That(t, err, test.ShouldBeNil)
+		test.That(t, isMoving, test.ShouldBeFalse)
+		test.That(t, arm1IsMovingCount.Load(), test.ShouldEqual, 1)
+	}
+
+	logs.TakeAll() // Reset logs.
+	resetAllIsMovingCounts()
+
+	{ // Duplicate local "fooArm" instances.
+		cfg := &config.Config{
+			Components: []resource.Config{
+				{
+					Name:       "fooArm",
+					Model:      model,
+					API:        arm.API,
+					Attributes: utils.AttributeMap{"armToRouteTo": "arm1"},
+				},
+				{
+					Name:       "fooArm",
+					Model:      model,
+					API:        arm.API,
+					Attributes: utils.AttributeMap{"armToRouteTo": "arm1"},
+				},
+			},
+		}
+		r.Reconfigure(ctx, cfg)
+
+		// Verify that a log was generated complaining about a collision.
+		test.That(t, logs.FilterMessageSnippet("collision").Len(), test.ShouldEqual, 1)
+
+		// Verify (refreshed) resource names returned to rClient contain _no_ "fooArm" instances.
+		expectedNames := []resource.Name{}
+		test.That(t, rClient.Refresh(ctx), test.ShouldBeNil)
+		actualNames := rClient.ResourceNames()
+		rdktestutils.VerifySameResourceNames(t, actualNames, expectedNames)
+
+		// Verify that "fooArm" can no longer be reached through ResourceByName from the rClient.
+		_, err := rClient.ResourceByName(arm.Named("fooArm"))
+		test.That(t, resource.IsNotFoundError(err), test.ShouldBeTrue)
+	}
+
+	logs.TakeAll() // Reset logs.
+	resetAllIsMovingCounts()
+
+	{ // One local "fooArm" instance and one remote "fooArm" instance.
+		cfg2 := &config.Config{
+			Components: []resource.Config{
+				{
+					Name:       "fooArm",
+					Model:      model,
+					API:        arm.API,
+					Attributes: utils.AttributeMap{"armToRouteTo": "arm2"},
+				},
+			},
+		}
+		r2.Reconfigure(ctx, cfg2)
+		cfg := &config.Config{
+			Components: []resource.Config{
+				{
+					Name:       "fooArm",
+					Model:      model,
+					API:        arm.API,
+					Attributes: utils.AttributeMap{"armToRouteTo": "arm1"},
+				},
+			},
+			Remotes: []config.Remote{
+				{
+					Name:    "r2",
+					Address: r2addr,
+				},
+			},
+		}
+		r.Reconfigure(ctx, cfg)
+
+		// Verify that a log was generated complaining about a collision.
+		test.That(t, logs.FilterMessageSnippet("collision").Len(), test.ShouldEqual, 1)
+
+		// Verify (refreshed) resource names returned to rClient contain only one "fooArm"
+		// instance (local).
+		expectedNames := []resource.Name{arm.Named("fooArm")}
+		test.That(t, rClient.Refresh(ctx), test.ShouldBeNil)
+		actualNames := rClient.ResourceNames()
+		rdktestutils.VerifySameResourceNames(t, actualNames, expectedNames)
+
+		// Call IsMoving() through "fooArm" from rClient, and assert that request routed to
+		// arm1.
+		fooArmResClient, err := rClient.ResourceByName(arm.Named("fooArm"))
+		test.That(t, err, test.ShouldBeNil)
+		fooArmClient, ok := fooArmResClient.(arm.Arm)
+		test.That(t, ok, test.ShouldBeTrue)
+		isMoving, err := fooArmClient.IsMoving(ctx)
+		test.That(t, err, test.ShouldBeNil)
+		test.That(t, isMoving, test.ShouldBeFalse)
+		test.That(t, arm1IsMovingCount.Load(), test.ShouldEqual, 1 /* increase */)
+		test.That(t, arm2IsMovingCount.Load(), test.ShouldEqual, 0 /* no increase */)
+
+		// Verify (refreshed) resource names returned to _r2Client_ also contains only one
+		// "fooArm" instance (local to r2, but remote to main machine).
+		expectedNames = []resource.Name{arm.Named("fooArm")}
+		test.That(t, r2Client.Refresh(ctx), test.ShouldBeNil)
+		actualNames = r2Client.ResourceNames()
+		rdktestutils.VerifySameResourceNames(t, actualNames, expectedNames)
+
+		// Call IsMoving() through "fooArm" from _r2Client_, and assert that request routed to
+		// arm2.
+		fooArmRes2Client, err := r2Client.ResourceByName(arm.Named("fooArm"))
+		test.That(t, err, test.ShouldBeNil)
+		fooArm2Client, ok := fooArmRes2Client.(arm.Arm)
+		test.That(t, ok, test.ShouldBeTrue)
+		isMoving, err = fooArm2Client.IsMoving(ctx)
+		test.That(t, err, test.ShouldBeNil)
+		test.That(t, isMoving, test.ShouldBeFalse)
+		test.That(t, arm1IsMovingCount.Load(), test.ShouldEqual, 1 /* no increase */)
+		test.That(t, arm2IsMovingCount.Load(), test.ShouldEqual, 1 /* increase */)
+
+		// Assert that prefixing the remote produces no new collision errors, and both arm1
+		// and arm2 are now reachable through rClient.
+		cfg.Remotes[0].Prefix = "r2"
+		r.Reconfigure(ctx, cfg)
+
+		test.That(t, logs.FilterMessageSnippet("collision").Len(), test.ShouldEqual, 1 /* no increase */)
+
+		expectedNames = []resource.Name{arm.Named("fooArm"), arm.Named("r2fooArm")}
+		test.That(t, rClient.Refresh(ctx), test.ShouldBeNil)
+		actualNames = rClient.ResourceNames()
+		rdktestutils.VerifySameResourceNames(t, actualNames, expectedNames)
+
+		// Assert that IsMoving calls route correctly.
+		isMoving, err = fooArmClient.IsMoving(ctx)
+		test.That(t, err, test.ShouldBeNil)
+		test.That(t, isMoving, test.ShouldBeFalse)
+		test.That(t, arm1IsMovingCount.Load(), test.ShouldEqual, 2 /* increase */)
+		test.That(t, arm2IsMovingCount.Load(), test.ShouldEqual, 1 /* no increase */)
+
+		r2fooArmResClient, err := rClient.ResourceByName(arm.Named("r2fooArm"))
+		test.That(t, err, test.ShouldBeNil)
+		r2fooArmClient, ok := r2fooArmResClient.(arm.Arm)
+		test.That(t, ok, test.ShouldBeTrue)
+		isMoving, err = r2fooArmClient.IsMoving(ctx)
+		test.That(t, err, test.ShouldBeNil)
+		test.That(t, isMoving, test.ShouldBeFalse)
+		test.That(t, arm1IsMovingCount.Load(), test.ShouldEqual, 2 /* no increase */)
+		test.That(t, arm2IsMovingCount.Load(), test.ShouldEqual, 2 /* increase */)
+	}
+
+	logs.TakeAll() // Reset logs.
+	resetAllIsMovingCounts()
+
+	{ // One local "fooArm" instance and two remote "fooArm" instances.
+		cfg3 := &config.Config{
+			Components: []resource.Config{
+				{
+					Name:       "fooArm",
+					Model:      model,
+					API:        arm.API,
+					Attributes: utils.AttributeMap{"armToRouteTo": "arm3"},
+				},
+			},
+		}
+		r3.Reconfigure(ctx, cfg3)
+		cfg2 := &config.Config{
+			Components: []resource.Config{
+				{
+					Name:       "fooArm",
+					Model:      model,
+					API:        arm.API,
+					Attributes: utils.AttributeMap{"armToRouteTo": "arm2"},
+				},
+			},
+		}
+		r2.Reconfigure(ctx, cfg2)
+		cfg := &config.Config{
+			Components: []resource.Config{
+				{
+					Name:       "fooArm",
+					Model:      model,
+					API:        arm.API,
+					Attributes: utils.AttributeMap{"armToRouteTo": "arm1"},
+				},
+			},
+			Remotes: []config.Remote{
+				{
+					Name:    "r2",
+					Address: r2addr,
+				},
+				{
+					Name:    "r3",
+					Address: r3addr,
+				},
+			},
+		}
+		r.Reconfigure(ctx, cfg)
+
+		// Verify that one new log was generated complaining about collision as we added r3.
+		test.That(t, logs.FilterMessageSnippet("collision").Len(), test.ShouldEqual, 1)
+
+		// Verify (refreshed) resource names returned to rClient contain only one "fooArm"
+		// instance (local).
+		expectedNames := []resource.Name{arm.Named("fooArm")}
+		test.That(t, rClient.Refresh(ctx), test.ShouldBeNil)
+		actualNames := rClient.ResourceNames()
+		rdktestutils.VerifySameResourceNames(t, actualNames, expectedNames)
+
+		// Call IsMoving() through "fooArm" from rClient, and assert that request routed to
+		// arm1 (local).
+		fooArmResClient, err := rClient.ResourceByName(arm.Named("fooArm"))
+		test.That(t, err, test.ShouldBeNil)
+		fooArmClient, ok := fooArmResClient.(arm.Arm)
+		test.That(t, ok, test.ShouldBeTrue)
+		isMoving, err := fooArmClient.IsMoving(ctx)
+		test.That(t, err, test.ShouldBeNil)
+		test.That(t, isMoving, test.ShouldBeFalse)
+		test.That(t, arm1IsMovingCount.Load(), test.ShouldEqual, 1 /* increase */)
+		test.That(t, arm2IsMovingCount.Load(), test.ShouldEqual, 0 /* no increase */)
+		test.That(t, arm3IsMovingCount.Load(), test.ShouldEqual, 0 /* no increase */)
+
+		// Assert that removing "fooArm" from the main machine leaves neither of the remote
+		// "fooArm"s accessible through the main machine due to name collision.
+		cfg.Components = nil
+		r.Reconfigure(ctx, cfg)
+
+		test.That(t, logs.FilterMessageSnippet("collision").Len(), test.ShouldEqual, 1 /* no increase */)
+
+		expectedNames = []resource.Name{}
+		test.That(t, rClient.Refresh(ctx), test.ShouldBeNil)
+		actualNames = rClient.ResourceNames()
+		rdktestutils.VerifySameResourceNames(t, actualNames, expectedNames)
+
+		// Verify that "fooArm" can no longer be reached through ResourceByName from the rClient.
+		_, err = rClient.ResourceByName(arm.Named("fooArm"))
+		test.That(t, resource.IsNotFoundError(err), test.ShouldBeTrue)
+
+		// Assert that prefixing one of the remotes allows access to both arm2 and arm3
+		// through the main machine.
+		cfg.Remotes[0].Prefix = "r2"
+		r.Reconfigure(ctx, cfg)
+
+		expectedNames = []resource.Name{arm.Named("fooArm"), arm.Named("r2fooArm")}
+		test.That(t, rClient.Refresh(ctx), test.ShouldBeNil)
+		actualNames = rClient.ResourceNames()
+		rdktestutils.VerifySameResourceNames(t, actualNames, expectedNames)
+
+		// Assert that IsMoving calls route correctly ("fooArm" to r3, and "r2fooArm" to r2).
+		fooArmResClient, err = rClient.ResourceByName(arm.Named("fooArm"))
+		test.That(t, err, test.ShouldBeNil)
+		fooArmClient, ok = fooArmResClient.(arm.Arm)
+		test.That(t, ok, test.ShouldBeTrue)
+		isMoving, err = fooArmClient.IsMoving(ctx)
+		test.That(t, err, test.ShouldBeNil)
+		test.That(t, isMoving, test.ShouldBeFalse)
+		test.That(t, arm1IsMovingCount.Load(), test.ShouldEqual, 1 /* no increase */)
+		test.That(t, arm2IsMovingCount.Load(), test.ShouldEqual, 0 /* no increase */)
+		test.That(t, arm3IsMovingCount.Load(), test.ShouldEqual, 1 /* increase */)
+
+		r2fooArmResClient, err := rClient.ResourceByName(arm.Named("r2fooArm"))
+		test.That(t, err, test.ShouldBeNil)
+		r2fooArmClient, ok := r2fooArmResClient.(arm.Arm)
+		test.That(t, ok, test.ShouldBeTrue)
+		isMoving, err = r2fooArmClient.IsMoving(ctx)
+		test.That(t, err, test.ShouldBeNil)
+		test.That(t, isMoving, test.ShouldBeFalse)
+		test.That(t, arm1IsMovingCount.Load(), test.ShouldEqual, 1 /* no increase */)
+		test.That(t, arm2IsMovingCount.Load(), test.ShouldEqual, 1 /* increase */)
+		test.That(t, arm3IsMovingCount.Load(), test.ShouldEqual, 1 /* no increase */)
+	}
+}
+
+func TestCrossAPINameCollision(t *testing.T) {
+	// Asserts machine-wide name uniqueness: two local resources may not share a simple name
+	// even under different APIs. On collision neither resource is built, and both become
+	// reachable again once the collision is resolved by renaming.
+	ctx := context.Background()
+
+	armModel := resource.DefaultModelFamily.WithModel(goutils.RandomAlphaString(8))
+	resource.RegisterComponent(arm.API, armModel, resource.Registration[arm.Arm, resource.NoNativeConfig]{
+		Constructor: func(context.Context, resource.Dependencies, resource.Config, logging.Logger) (arm.Arm, error) {
+			return &inject.Arm{}, nil
+		},
+	})
+	defer resource.Deregister(arm.API, armModel)
+
+	sensorModel := resource.DefaultModelFamily.WithModel(goutils.RandomAlphaString(8))
+	resource.RegisterComponent(sensor.API, sensorModel, resource.Registration[sensor.Sensor, resource.NoNativeConfig]{
+		Constructor: func(context.Context, resource.Dependencies, resource.Config, logging.Logger) (sensor.Sensor, error) {
+			return &inject.Sensor{}, nil
+		},
+	})
+	defer resource.Deregister(sensor.API, sensorModel)
+
+	logger, logs := logging.NewObservedTestLogger(t)
+
+	// Start with an arm and a sensor that share the name "dup" across different APIs.
+	cfg := &config.Config{
+		Components: []resource.Config{
+			{Name: "dup", Model: armModel, API: arm.API},
+			{Name: "dup", Model: sensorModel, API: sensor.API},
+		},
+	}
+	r := setupLocalRobot(t, ctx, cfg, logger)
+
+	// A collision was logged, and neither resource is reachable.
+	test.That(t, logs.FilterMessageSnippet("collision").Len(), test.ShouldBeGreaterThan, 0)
+	rdktestutils.VerifySameResourceNames(t, r.ResourceNames(), []resource.Name{})
+	_, err := r.ResourceByName(arm.Named("dup"))
+	test.That(t, resource.IsNotFoundError(err), test.ShouldBeTrue)
+	_, err = r.ResourceByName(sensor.Named("dup"))
+	test.That(t, resource.IsNotFoundError(err), test.ShouldBeTrue)
+
+	// Resolve the collision by renaming: both resources should now be reachable.
+	cfg = &config.Config{
+		Components: []resource.Config{
+			{Name: "armDup", Model: armModel, API: arm.API},
+			{Name: "sensorDup", Model: sensorModel, API: sensor.API},
+		},
+	}
+	r.Reconfigure(ctx, cfg)
+
+	rdktestutils.VerifySameResourceNames(t, r.ResourceNames(),
+		[]resource.Name{arm.Named("armDup"), sensor.Named("sensorDup")})
+	_, err = r.ResourceByName(arm.Named("armDup"))
+	test.That(t, err, test.ShouldBeNil)
+	_, err = r.ResourceByName(sensor.Named("sensorDup"))
+	test.That(t, err, test.ShouldBeNil)
+}
+
+func TestThreeWayNameCollision(t *testing.T) {
+	// Three local resources share a name across APIs. The first two collide (the first
+	// is marked for removal, the second refused); the third must NOT slip through by
+	// finding only the marked-for-removal sibling. None of the three may be reachable.
+	ctx := context.Background()
+
+	armModel := resource.DefaultModelFamily.WithModel(goutils.RandomAlphaString(8))
+	resource.RegisterComponent(arm.API, armModel, resource.Registration[arm.Arm, resource.NoNativeConfig]{
+		Constructor: func(context.Context, resource.Dependencies, resource.Config, logging.Logger) (arm.Arm, error) {
+			return &inject.Arm{}, nil
+		},
+	})
+	defer resource.Deregister(arm.API, armModel)
+
+	sensorModel := resource.DefaultModelFamily.WithModel(goutils.RandomAlphaString(8))
+	resource.RegisterComponent(sensor.API, sensorModel, resource.Registration[sensor.Sensor, resource.NoNativeConfig]{
+		Constructor: func(context.Context, resource.Dependencies, resource.Config, logging.Logger) (sensor.Sensor, error) {
+			return &inject.Sensor{}, nil
+		},
+	})
+	defer resource.Deregister(sensor.API, sensorModel)
+
+	msModel := resource.DefaultModelFamily.WithModel(goutils.RandomAlphaString(8))
+	resource.RegisterComponent(movementsensor.API, msModel,
+		resource.Registration[movementsensor.MovementSensor, resource.NoNativeConfig]{
+			Constructor: func(context.Context, resource.Dependencies, resource.Config, logging.Logger) (movementsensor.MovementSensor, error) {
+				return &inject.MovementSensor{}, nil
+			},
+		})
+	defer resource.Deregister(movementsensor.API, msModel)
+
+	logger := logging.NewTestLogger(t)
+	r := setupLocalRobot(t, ctx, &config.Config{
+		Components: []resource.Config{
+			{Name: "dup", Model: armModel, API: arm.API},
+			{Name: "dup", Model: sensorModel, API: sensor.API},
+			{Name: "dup", Model: msModel, API: movementsensor.API},
+		},
+	}, logger)
+
+	rdktestutils.VerifySameResourceNames(t, r.ResourceNames(), []resource.Name{})
+	for _, n := range []resource.Name{arm.Named("dup"), sensor.Named("dup"), movementsensor.Named("dup")} {
+		_, err := r.ResourceByName(n)
+		test.That(t, resource.IsNotFoundError(err), test.ShouldBeTrue)
+	}
+}
+
+func TestUserBuiltinNameCollision(t *testing.T) {
+	// The default services legitimately share the "builtin" name across service APIs and are exempt,
+	// so a lone non-default resource may also use "builtin". But two non-default resources sharing
+	// "builtin" across APIs collide like any other duplicated name: neither is built and it is
+	// logged. Renaming one resolves the collision and the surviving lone "builtin" is reachable.
+	ctx := context.Background()
+
+	armModel := resource.DefaultModelFamily.WithModel(goutils.RandomAlphaString(8))
+	resource.RegisterComponent(arm.API, armModel, resource.Registration[arm.Arm, resource.NoNativeConfig]{
+		Constructor: func(context.Context, resource.Dependencies, resource.Config, logging.Logger) (arm.Arm, error) {
+			return &inject.Arm{}, nil
+		},
+	})
+	defer resource.Deregister(arm.API, armModel)
+
+	sensorModel := resource.DefaultModelFamily.WithModel(goutils.RandomAlphaString(8))
+	resource.RegisterComponent(sensor.API, sensorModel, resource.Registration[sensor.Sensor, resource.NoNativeConfig]{
+		Constructor: func(context.Context, resource.Dependencies, resource.Config, logging.Logger) (sensor.Sensor, error) {
+			return &inject.Sensor{}, nil
+		},
+	})
+	defer resource.Deregister(sensor.API, sensorModel)
+
+	logger, logs := logging.NewObservedTestLogger(t)
+
+	// An arm and a sensor both named "builtin": a cross-API collision, neither reachable.
+	r := setupLocalRobot(t, ctx, &config.Config{
+		Components: []resource.Config{
+			{Name: resource.DefaultServiceName, Model: armModel, API: arm.API},
+			{Name: resource.DefaultServiceName, Model: sensorModel, API: sensor.API},
+		},
+	}, logger)
+
+	test.That(t, logs.FilterMessageSnippet("collision").Len(), test.ShouldBeGreaterThan, 0)
+	_, err := r.ResourceByName(arm.Named(resource.DefaultServiceName))
+	test.That(t, resource.IsNotFoundError(err), test.ShouldBeTrue)
+	_, err = r.ResourceByName(sensor.Named(resource.DefaultServiceName))
+	test.That(t, resource.IsNotFoundError(err), test.ShouldBeTrue)
+
+	// Renaming the sensor off "builtin" resolves the collision. The lone remaining "builtin" arm is
+	// then allowed (it does not collide with the exempt default services).
+	r.Reconfigure(ctx, &config.Config{
+		Components: []resource.Config{
+			{Name: resource.DefaultServiceName, Model: armModel, API: arm.API},
+			{Name: "sensorOK", Model: sensorModel, API: sensor.API},
+		},
+	})
+	_, err = r.ResourceByName(arm.Named(resource.DefaultServiceName))
+	test.That(t, err, test.ShouldBeNil)
+	_, err = r.ResourceByName(sensor.Named("sensorOK"))
+	test.That(t, err, test.ShouldBeNil)
+}
+
+func TestCrossAPIRemoteNameCollision(t *testing.T) {
+	// A remote resource whose name collides with a local resource of a different API is hidden
+	// from the machine's resource list (local wins) and logged, but is still directly resolvable
+	// by its exact name+API.
+	ctx := context.Background()
+
+	armModel := resource.DefaultModelFamily.WithModel(goutils.RandomAlphaString(8))
+	resource.RegisterComponent(arm.API, armModel, resource.Registration[arm.Arm, resource.NoNativeConfig]{
+		Constructor: func(context.Context, resource.Dependencies, resource.Config, logging.Logger) (arm.Arm, error) {
+			return &inject.Arm{}, nil
+		},
+	})
+	defer resource.Deregister(arm.API, armModel)
+
+	sensorModel := resource.DefaultModelFamily.WithModel(goutils.RandomAlphaString(8))
+	resource.RegisterComponent(sensor.API, sensorModel, resource.Registration[sensor.Sensor, resource.NoNativeConfig]{
+		Constructor: func(context.Context, resource.Dependencies, resource.Config, logging.Logger) (sensor.Sensor, error) {
+			return &inject.Sensor{}, nil
+		},
+	})
+	defer resource.Deregister(sensor.API, sensorModel)
+
+	logger, logs := logging.NewObservedTestLogger(t)
+
+	// Remote exposes a sensor named "shared".
+	remote := setupLocalRobot(t, ctx, &config.Config{
+		Components: []resource.Config{{Name: "shared", Model: sensorModel, API: sensor.API}},
+	}, logger.Sublogger("remote"))
+	options, _, addr := robottestutils.CreateBaseOptionsAndListener(t)
+	test.That(t, remote.StartWeb(ctx, options), test.ShouldBeNil)
+
+	// Main has a local arm named "shared" and connects to the remote with no prefix.
+	main := setupLocalRobot(t, ctx, &config.Config{
+		Components: []resource.Config{{Name: "shared", Model: armModel, API: arm.API}},
+		Remotes:    []config.Remote{{Name: "r", Address: addr}},
+	}, logger.Sublogger("main"))
+
+	// Local arm "shared" wins; the remote sensor "shared" is hidden from the resource list, and
+	// the cross-API collision is logged.
+	testutils.WaitForAssertionWithSleep(t, time.Millisecond*100, 300, func(tb testing.TB) {
+		names := main.ResourceNames()
+		test.That(tb, names, test.ShouldContain, arm.Named("shared"))
+		test.That(tb, names, test.ShouldNotContain, sensor.Named("shared"))
+		test.That(tb, logs.FilterMessageSnippet("collision with a remote resource").Len(), test.ShouldBeGreaterThan, 0)
+	})
+
+	// The local arm "shared" is reachable.
+	_, err := main.ResourceByName(arm.Named("shared"))
+	test.That(t, err, test.ShouldBeNil)
+
+	// The hidden remote sensor "shared" is still directly resolvable by exact name+API.
+	res, err := main.ResourceByName(sensor.Named("shared"))
+	test.That(t, err, test.ShouldBeNil)
+	_, ok := res.(sensor.Sensor)
+	test.That(t, ok, test.ShouldBeTrue)
+}
